Run the keepalived exporter as non-root
The capabilities are still quite extensive, but it is a start.
Maybe we can also reduce the privileges of the keepalived process
itself at some point so that we can get rid of CAP_KILL and
CAP_DAC_READ_SEARCH.

The remaining capabilities grant the following privileges
(according to capabilities(7)):

* interface configuration;
* administration of IP firewall, masquerading, and accounting;
* modify routing tables;
* bind to any address for transparent proxying;
* set type-of-service (TOS)
* clear driver statistics;
* set promiscuous mode;
* enabling multicasting;
* use setsockopt(2) to set the following socket options:
  SO_DEBUG, SO_MARK, SO_PRIORITY (for a priority outside the range
  0 to 6), SO_RCVBUFFORCE, and SO_SNDBUFFORCE.

* Bypass file read permission checks and directory read and execute
  permission checks;
* invoke open_by_handle_at(2);
* use the linkat(2) AT_EMPTY_PATH flag to create a link to a file
  referred to by a file descriptor.

* Bypass permission checks for sending signals (see kill(2)).
  This includes use of the ioctl(2) KDSIGACCEPT operation.

This is still quite extensive (and potentially destructive), but
most notably it doesn’t include arbitrary filesystem writes
(unless by escalating using one of the given privileges).