Implement support for kube-router
[kube-router][1] is an alternative Cluster Networking
implementation for kuberentes. It uses the Brigde CNI to create
a very performant and feature rich networking setup for
Kubernetes.

Flannel on the other hand uses an overlay network on top of the
existing network between the workers. The upside of this is that
it requires virtually zero configuration and will work nearly
everywhere out of the box. The downside is that it uses a lower
MTU and [is slow][2].

This commit provides optional support for kube-router. The default
networking setup will still be flannel in order to support
existing clusters. For new clusters we highly recommend using
kube-router.

Compared to flannel, kube-router provides us with:

- Easy integration of non-k8s machines in the cluster network
- Support for Network Policies
- Improved throughput (factor 2 to 4), which equates improved
  storage performance (one test has shown a factor of 2 on read
  throughput)
- Possible future improvements to the LBaaS integration, which
  reduce latency and improve throughput

MIGRATION

To migrate from flannel to kube-router in a living cluster, a
downtime announcement is strictly required. All pods need to be
restarted. The migration is triggered by setting

    k8s_network_plugin = "kube-router"
    k8s_network_plugin_switch_restart_all_namespaces = true or false

in the cluster configuration and rolling out the new configuration
in the cluster. Please see the documentation of the options in
jenkins/config.template.toml for more details.

Fixes #54.

   [1]: https://github.com/cloudnativelabs/kube-router
   [2]: https://gitlab.cloudandheat.com/lcm/managed-k8s/-/issues/54#note_77510