Provide init.sh action
This action is supposed to bootstrap a cluster repository without
executing a full rollout like manage_cluster.sh does.

It manages the .gitignore file and ensures presence and correct
configuration of all submodules. In an empty repository, it will
also create and bootstrap the `config/config.toml` (while leaving
the existing config untouched).

This is in context of #68 to enforce a proper gitignore.