Use apply: yes
This uses openshift.dynamic.apply instead of the default merge
strategy. Much like kubectl apply, this keeps information about
the last applied state in the resource itself, allowing to do a
true three-way merge instead of a dumb dict merge.

This has the advantage that annotations, labels and other
properties which were *deleted* instead of changed or added can
be handled correctly.

Note that this change will only help with the *next* update to the
resources, since the metadata is only written when `apply: yes` is
already set.

Fixes #120.