Introduce resource requests and limits for Rook Pods
Especially with the OSDs, this helps to avoid OOM situations
because they are now configured based on the requests/limits set
in this configuration. Ceph contains code to detect the
environment variables set by Rook which contain the
limits/requests and will configure its internal memory use limits
and cache sizes accordingly.

In context of #119. Fixes #122.