Introduce resource requests and limits for monitoring
This helps kubernetes to make smart decisions about the placement
of the pods for the monitoring services.

In contrast to Rook, the monitoring pods are not auto-scaling
based on the limits set from k8s, so we have to tread more
carefully here around which numbers we give them.

Especially Prometheus in version 2.x does not allow any control
over the used memory anymore; it makes extensive used of memory
mapped files to delegate that to the OS. While the advantage is
that it’ll auto-scale to a certain extent based on the memory
manager of the OS, the downside is that we cannot assess the real
*required* memory use by just looking at the RSS of a running
prometheus process. We’ll have to gather experience here.

In context of #119.