Restructure CI cluster for new resource limits
* We now use 3xL + 3xM as workers instead of 6xM
* Rook and Monitoring services are separated; Rook runs on the L
  workers, Monitoring runs on the M workers.
* Monitoring is not tainted, i.e. the tests can run on the M
  workers, too (we would normally not do that in production)

The advantage of this layout is that Rook and Monitoring cannot
steal each other resources due to bad packing of the Pods, making
the cluster construction hopefully more deterministic.

The raise in resources is needed because the cluster was too
thinly provisioned before; the Absolute Minimum Ratings for Rook
Ceph alone require 18 GB of RAM (spread across three workers,
that’s 6GB of RAM; That is without the derating required to
account for supporting services (such as kube-router and
node-exporter)).