Make any pre-flight check failure abort the run for all nodes
If any pre-flight check fails, then the cluster is in no state to
be upgraded. This means that we have to abort immediately. Without
this, a single node whose preflight check passes may continue into
the upgrade, with possibly devastating effects if the upgrade has
not started yet and a non-master node is the only node passing.