Re-order upgrade steps so that the drain happens first
This is important to be able to resume the upgrade in case the
drain fails for any reason (for example, stubborness of kubectl
drain). Previously, this would lead to a node which has to be
fixed manually, since the kube{let,adm,ctl} packages have already
been upgraded. With this patch, the upgrade happens only after
the drain was successful.

The downside of this approach is that the node will stay longer
in drained + cordoned state.