Support resumption of interrupted upgrades
By allowing situations where the installed version matches the
target version, we allow resuming an upgrade which was
interrupted between two nodes without having to resort to use of
the ansible-playbook option --limit; that makes the resumption
safer.

Note that this does not allow resuming a partial upgrade on a
single node; if an upgrade gets interrupted on a node after the
first package has been upgraded, manual operator intervention is
still necessary. However, after that is done, the upgrade can now
be resumed with ansible.

In addition, this makes this playbook idempotent.