Fix disabling of rook
Due to the precedence of variables in ansible, the playbook group
vars win against the group vars given in the inventory. Setting
rook to enabled in the group vars in the playbook thus forces rook
to be deployed in all clusters.

This is not desirable; for example, the future monitoring cluster
is not supposed to use rook.

Fixes #140.