Add option to disable AZ management in Terraform
If the flag is set to false (the default is true), Terraform will
not request a specific availability zone for the instances it
creates.

This can be useful in situations where we either don’t care about
the AZ (e.g. in CI) or where the cloud can currently not fulfill
the AZ requirements we would have (e.g. if a rack is down or
something like that).