Provide a route into the ClusterIP network from the gateways
Note the FIXME inside that. This route breaks down as soon as the
first master goes down.

This is annoying, but there’s no easy workaround. We cannot use
Equal Cost Multi Path (ECMP) routing because that’d break with the
stateful DNAT happening at the ClusterIP -> Pod transition on the
k8s nodes. We also don’t have an easy way to tell bird to use the
BGP up-ness to determine the nexthop.

The best way would be if the BGP peers also announced the ClusterIP
network; then bird could pick one of them as nexthop. But that
doesn’t seem to be feasible at the moment.