Make the IP networks used by Kubernetes configurable
This is a requirement for the global monitoring kubernetes, where
we need to interface with all other clusters. The global monitoring
cluster cannot use the same Pod or Service IP space, otherwise
it’s not possible to set up even a Layer 7 proxy without extensive
effort.

(Although not completely impossible if a protocol break such as
via a Unix domain socket is used :-X.)