Enable automatic IP address management for per-cluster WireGuard users
This makes toml_helper.py not require that an IP address is given
in config.toml for a user. It will auto-allocate an IP address in
that case like for any other user.

Note that the allocated IP address is not written back to
config.toml but, like the other addresses, to wireguard_ipam.toml.