Refactor the lookup for wireguard user to use toml
See issue #151

The "wg-up.sh" script was using jq to parse "config/config.json" to look
for the defined users in the wireguard setup. An error with `ARGS` not
being defined was due to a syntax error, where `$user` should have been
leveraged to use the given "user" parameter. That is what triggered the
jq error.

Anyway, the file "config/config.json" does not exist, and the existence
of the user should be checked by looking at
"config/wireguard_ipam.toml". As it is a toml file and there exists no
equivalent to jq for toml, a Python command is now used to check the
presence of the user in the file.