Add validation for the Kubernetes resources applied
Several Kubernetes resources are created on the Kubernetes cluster, for
monitoring for instance. These resources are created using the
Kubernetes module of Ansible. A validation is now performed on the
manifests before the actual creation, to ensure that all field have the
right names and content. The resources which did not pass the validation
were modified accordingly.