Change kube_thanos include to include_tasks
When using `include`, the task included was somehow loaded, even if
skipped. This, in turn, executed the jinja in the task definition. But
in one template one lookup is performed on a file that does not exist,
an error is raised, even if the task is skipped.

In this case, a file is not created if thanos is not enabled. Thus the
role failed. When using `include_tasks`, the task that holds the lookup
is actually skipped.

[skip ci]