Add IPv6 support for the toml_helper
The toml_helper.py is able to support IPv4, IPv6 and IPv6 and IPv4
addresses for users in the wireguard_ipam.toml file.
If the config wireguard_ipam.toml already exists and a second IP range
with a different version is introduced, the script keeps the old IP
address and adds a new one from the new IP range.
The IPv4 and IPv6 address are handled separately in its own fields,
a list is not necessary, because a user cannot have more
than one IP address of an address family.