mon: expose alertmanager's secret name to config
The purpose of this commit is to give the customer an opportunity to use a
configuration secret for alertmanager that is not controlled by
kube-prometheus and hence not overwritten on a roll-out. To use it,
override the value of `monitoring_alertmanager_config_secret` in
config.toml  to, e.g., `am-customer-config` and run stage3. The
alertmanager-CRD will then adapt the auto-generated statefulset to use
that secret. Of course the pods will fail to start if the secret does
not exist.

I checked that `rebuild` does not generate a default secret with the
custom name we've chosen above.