kubeadm: set node name during 'init'
cloud-init is responsible of setting the hostname of a node. On CentOS
the FQDN is used, e.g., `mynode.novalocal`. `kubeadm init` uses the
output of `hostname -f` which leads to a mismatch between the hostname
in the ansible inventory and the name of the node in k8s. By adding the
flag `--node-name` in `kubeadm init` we ensure that the names match.

(Strangely, other nodes than the master-0 are not registered by their
FQDN.)