Bird: Remove hardcoded route to ClusterIP network
Currently the ClusterIP network is made available
from the gateway via a static route hardcoding the
first master as next hop. Thus, reachability to
the ClusterIP network is disrupted when the first
master node goes down.

As we don't currently use the ClusterIP network,
this commit removes the static route in order to
avoid problems in the long run in case we start
using it at some point in the future but forget to
add proper failover.

(Better having to fix it during development than
having it blow up during production use.)