Remove unattended-upgrades for now
This effectively disables automatic (security) updates on Ubuntu nodes.
unattended-upgrades is removed because system updates can be highly
disruptive, e.g., when the docker daemon is affected and its restart
also causes a restart of all containers. One way around this would be to
blacklist certain services from updates like docker. The question is
where to draw the line in this case.