kubeadm:support ServiceAccountToken vol projection
This commit adds two flags to the kube-apiserver that are necessary for
ServiceAccountToken volume projection to work correctly. The projection
is needed, e.g., by istio to create trustworthy JWT tokens. Existing
clusters are not affected by this change. Should they also require this
feature, one can quickly add the flags to
/etc/kubernetes/manifests/kube-apiserver.yaml. The container is
restarted automatically on a change.