ipsec: Implement support for multiple networks on both sides
This allows for a full site-to-site setup. IPsec routes to the
remote are imported by bird and forwarded to the k8s cluster so
that pods can reach there. nftables is configured to allow
forwarding (and it’ll also SNAT the traffic... we should think
about whether we want that).