ipsec: Do not masquerade traffic from remote peers
This commit adds a rule to the postrouting chain in the nat table
(nftables) that skips masquerading for remote ipsec peers. Hereby we
prevent unnecessary SNAT in the scenario R --IPsec--> Pod. We probably
don't want to expose the pod and service subnet to the outside, though.
I.e., the SNAT for Pod --IPsec--> R remains in place.