Mark update of gateway as disruptive action
An update of the gateway implies a reboot. Although we have a fail-over
mechanism with keepalived in place, the disruption can be and is noted
by the user. All connections have to be restarted twice when the master
role is re-assigned (gw1 -> gw2 -> gw1). Therefore, the following
changes are introduced:

- There's an additional state `uninitialized` for the gateways. The
  availability of keepalived, haproxy and wireguard acts as a condition.
  Changes to `uninialized` nodes are assumed to be not disruptive, hence
  they're applied in parallel.
- `Initialized` gateways aren't updated unless the kraken is released.