Allow spawning clusters without monitoring enabled
Ansible will look for the files to evaluate `loop` even if the
tasks will be skipped thanks to a failing `when`; this makes
sense because when is evaluated after `loop`.

To avoid issues from there, we modify the error handling for
the cases where the `when` will evaluate to false. In effect,
this allows to spawn clusters with monitoring set to false.