ipsec: run strongswan only on VRRP Master
I did some tests with the current setup (basically cross-pinged the
gateways in two clusters). It turned out that it is a bad idea to have
multiple strongswan instances active. Consider the following example:

We assume that gw1 and gw1' are VRRP masters in their designated
clusters and hence hold the IP that's configured as remote endpoint in
strongswan. gw2 now wants to ping gw2'.

(1) gw2 -> gw1' (gw2 establishes a tunnel with gw1' and sends an ICMP
packet)
(2) gw1' forwards the decrypted packet to gw2'.
(3) gw2' -> gw1 (the return path is different)
(4) gw1 -> gw2

This does not only cause a knot in my brain but is also annoying because
the tunnel is stateful. gw1 and gw1' have each established a tunnel into
the other cluster but not with the public endpoint. Traffic is not
forwarded because of an authentication mismatch and things have to be
restarted.