ipsec/bird: exchange routes with VRRP backups
swanctl/some other IPsec magic adds a route to the remote
peer on the localhost. By disabling swanctl on the VRRP backup gateways
they obviously cannot reach the other side. This commit tells bird to
communicate the route to the remote site with the VRRP master as next
hop.

Note that this is a cosmetic change because from a practical perspective
there is no need for a VRRP backup gateway to reach the other side.