journald: store logs persistently and limit size
Not storing logs persistently causes us to not being able to
investigate issues when they occur, especially on the gateways.
On the gateways, ephemeral logs will not occupy more than 10MB,
which is nothing, it doesn’t even last for 24h.

By moving logs to persistent storage, we of course increase
I/O load (especially potentially considerable on workers), which
is why this can still be turned off and changed for k8s workers.

The default limits are 4GB or 10%, whichever is less. It will
still keep a certain amount of storage free.