ipsec: don't stop strongswan-swanctl on rollout
The swanctl service was stopped on all nodes after a rollout to allow
keepalived to start the service only on the master. This only worked
when keepalived was always restarted to make it call the notify script.
Nowadays the keepalived restart is not mandatory, hence swanctl was
stopped but not (re)started, until there was a change in keepalived.
This commit
- keeps swanctl running on the VRRP master
- restarts swanctl if one of its config files changes