Setup Calico Resources on Master-Nodes
Setting up the basis for the calico networking.
We install calicoctl, configure secrets for the TLS encryption,
setup the RBAC as well as PSP support and configure the
underlying OS to use calico as CNI plugin.
Basically this does the preparation steps according
to the calico the hard way tutorial.

Issue: #218