Move calico related BGP conf to k8s-bgb
If the cluster is created with kube-router as CNI plugin,
stage 3 will fail cause the calico BGP conf is rolled out
in addition to the kube-router conf.

The calico and kubre-router BGP conf are nearly identical,
and restarting the bird service leads to a "Symbol already
defined" error. Therefore, the calico BGP conf roll out is
adjusted to the new circumstances and rolled out only if calico
is chosen as CNI plugin.

Short: prevent duplicate config for bird

Issue: #218