kube-router: restart container on loss of routes
This commit extends the livenessProbe of the kube-router daemonsets
which now compares the availibilty of routes on a worker to the routes
offered by its bgp peers. On a mismatch, i.e., if the worker lacks one
or more routes into the k8s network the pod is restarted.

Context: External events like an expired DHCP lease or a restart of
systemd-networkd during a system update will cause networkd to set its
interfaces down and up. Consequently the routes are dropped, too.
kube-router does not re-sync routes on a node automatically but has to
be restarted.

This commit also moves the AS numbers to the k8s-config role.
Note that this is mostly a best-effort workaround to cope with a bug in
kube-router.

Notes on the expression of the livenessProbe:

- entries with `kube-bridge` (i.e., routes into the node-local pod/service network)
are excluded because the bridge IP is set lazily by the CNI plugin.
Nodes without any pods inside the pod network don't have an IP
configured on the bridge and hence no route which would cause a
mismatch
- using `gobgp neigh` compared to `gobgp global rib` has the advantage
of only showing routes of other peers, not our own route (see in
conjunction with the previous point)

Related:

- https://github.com/systemd/systemd/issues/13003 ("systemd-networkd
  removes IPv4 address during DHCP renewal")
- https://redmine.cloudandheat.com/issues/28655 (opsi ticket)
- https://github.com/cloudnativelabs/kube-router/issues/509 (issue
  report that kube-router does not re-sync routes)