Add DualStack support to wireguard
The wg-iup.sh script has been updated to support IPv6 addresses.
The wireguard.conf has been modified to enable IPv6 forwarding if
DualStack is enabled. The client and server configuration templates
are adjusted to add IPv6 addresses and prefix if DualStack support
is enabled.

For the connection to cluster nodes, it is necessary that radvd is
configured on the gateways.

Issue: #227