Add DualStack support to bird and k8s-bgp
If DualStack is enabled, we need additional configs for bird
on the gateway nodes such that it imports IPv6 routes propagated
by the master nodes. To do so, bird6 needs to be enabled. An
additional dropdown directory is created so that the k8s-bgp role
can dynamically add the peering with the k8s nodes via IPv6.

Issue: #227