Add DualStack support to terraform
Create IPv6 related resources based on the newly introduced
dualstack_support variable. This includes an IPv6 subnet as well
as an IPv6 router interface. In addition, the outputs are adjusted
such that the DualStack support can be setup by Ansible.

Each node gets an IPv4 as well as an IPv6 address from the specified
subnet. The IPv6 is created dynamically based on the value of the
dualstack_support variable.

We do use dhcpv6-stateful, because SLAAC does not work with subnets
smaller than /64. In addition, some security guides reject the use of
SLAAC or DHCP stateless, because it allows malicious actors on the
same layer 2 to cause more havoc than they usually could.

These tasks do preserve compatibility with the IPv4 only setup.

Issue: #227