Install radvd on gateways
Radvd propagates the correct route back to the wireguard client.

The Router Advertisement Daemon needs to be installed and enabled
on the gateways for the wireguard IPv6 support. Radvd sends router
advertisement messages. This allows to automatically configure
addresses, settings, routes, and choose a default router.

The gateway nodes must have IPv6 forwarding enabled.

Radvd is necessary on the gateways to enable wireguard functionality.
Without radvd it is possible to route into the cluster, but host nodes
do not know how to route packets back.

Radvd is configured in such a way, that only the gateway that holds the
virtual IPv6 address sends router advertisements. The VIP is managed
by keepalived. With the combination of radvd and keepalived, the
host nodes know the route back to the correct gateway.

Issue: #227