Add DualStack support to k8s components
To enable IPv4/IPv6 dual-stack support, the IPv6DualStack feature gate
has to be enabled for kube-apiserver, kube-controller-manager,
kubelet, and kube-proxy. In addition, it is necessary that kube-proxy
runs in IPVS mode. When kubeadm initializes, it will take care, that
these settings are rolled out.

Other dual-stack related settings are currently not supported well by
kubeadm. There is a pull request which adds configuration options
related to dual-stack support to the kubeadm config. Please refer to
https://github.com/kubernetes/kubeadm/issues/1612

Issue: #227