Add DualStack support to calico
For DualStack support, the CNI config has to be adjusted such that
Calico's IPAM will allocate both IPv4 and IPv6 addresses for each new
Pod or Service. In addition, the env variables for calico/node have to
be adjusted such that calico/node can autodetect the host nodes IPv6
address and use it in its BGP IPv6 configuration.
IPv6 support for Felix has to be enabled as well, so that Felix knows
to program routing and iptables for IPv6 as well as for IPv4.

Adjusting the check-calico role:
With Dual-Stack enabled, Pods get an IPv4 as well as an IPv6 address.
If this is the case, check that Pods can speak to each other using
the IPv6 addresses. Ansible tries to expand variables before skipping
the task. This leads to an undefined variable error if only one IP is
assigned to a Pod.
This commit adds a default value if no second IP is defined, so the
task can be properly skipped.

This commit preserves compatibility with the IPv4 only setup.

Issue: #227

Add task that checks IPv6 connectivity

Issue: #227