Adjust scripts for bare-metal deployment
Current scripts are depending on a Wireguard installation, which is not
needed for a bare-metal / standalone deployment.
Also certain task not to be skipped during
deployment, as they are only needed for full (on OpenStack)
deployments with gateway nodes.

- init.sh
Added a rudimentary access checks for C&H wg_user, passwordstore,
ch-role-users repos, which allows non-C&H users to also create their
cluster repo without editing any scripts.
As Terraform is not used by this type of deployment, the inventory
folder and a host template has to be created.

- lib.sh
wg_user folder dependency has been added, which skips the
initialization of Wireguard variables for bare-metal / standalone
deployments.
Also several folder and file variables have been added, along a
function for the repo availability check for init.sh

- stage2.sh / stage3.sh
Had to be adjusted so that it copies, based on folder or file
availability, the hosts inventory file from stage2 to stage3, including
the various other config variables config files, which are needed
during stage3 execution.
Also checking for the wg_user folder in the cluster repo, if not
available establishment of Wireguard tunnel will be skipped.

- toml_helper.py
Also added a wg_user folder/repo availability check, in order to be
executable on bare-metal / standalone deployments or for non-C&H users.