Split services into separate ansible directories
This introduces two new ansible playbooks:

1. "service layer", which introduces APIs and k8s features which
   are not part of a vanilla k8s cluster but which may still be
   useful/commonly used.

   This includes cert manager, ingress as well as the Prometheus
   custom resource API.

2. "managed services", which includes the value-added services
   of our managed-k8s product. This includes a Prometheus
   deployment as well as the rook-based storage cluster.

During this split, a few cross-layer dependencies had to be
resolved, which has practical effects on clusters. The following
changes were made:

- Instead of the previously rather flexible way of specifying
  separate tolerations and node labels as targets for services,
  this is now reduced to the ["Scheduling Key"][1] concept which
  has been introduced in Yaook.

  This simplifies configuration significantly.

- Storage node plugins can now run with tolerations for all
  taints, with a single toleration or no toleration at all.
  There is no default toleration tailored for the mk8s control
  plane anymore.

  This is to avoid a dependency on the scheduling key (taint)
  defined for managed services in the earlier stages.

This has direct implications on configuration files, please see
the changes in the template config for details.

To carry configuration options between the layers, the "legacy"
layer (in `ansible`) will export JSON files with a subset of the
options provided. Those JSON files will be read as extra vars
in the subsequent layers.

This is a hack at best and needs to be resolved in the future,
possibly in accordance with #113.

   [1]: https://docs.yaook.cloud/concepts/scheduling.html