Use handlers to restart and configure docker
Rather than registering the output of the docker installation and
executing the configuration and restart depending on the outcome,
this adjusts the role to use handlers for that.
This satisfies ansible linting.

At the end of the role, a flush handlers is introduced so that the
docker daemon gets configured and restarted immediately if necessary.

Issue: #242