Adjust remove flannel logic
This commit introduces a handler to restart kubelet.
This satisfies ansible linting.

This commit changes the conditionals which were used to remove flannel.
Before flannel resources get removed, we have a task that explicitly
fails if the removal is not permitted. Therefore, we can simplify the
logic of the subsequent tasks.

Issue: #242