Adjust calico CI job to use DualStack setup
We need to manipulate the Terraform variables so that Terraform will
prepare DualStack resources (which is disabled by default).

Terraform supports setting environment variables to set terraform
variables. However, the environment variable do NOT override the
value of the config.tfvars.json!

This commit introduces an environment variable TF_VAR_dualstack_support
which overrides the default value of the dualstack_support variable.

Issue: #227