Fix toml helper DualStack-Support
When adding a user via the config file, there are already IP addresses
defined for that user. The toml_helper did not properly check these
assigned addresses and outputs the IPv4 address as the IPv6 address.

So the user gets the IPv4 address assigned as IPv4 and IPv6 address.
This causes Wireguard to fail, because the used IP is duplicated.

The error happens, because of a copy-paste typo.

This commit fixes the typo.

Issue: #227