Remove x-prefix in comparisons
Some older shells would get confused if the first argument started with
a dash, or consisted of ! or (. As a workaround, people would prefix
variables and values to be compared with x to ensure the left-hand side
always started with an alphanumeric character.

POSIX ensures this is not necessary, and all modern shells now follow
suite.

https://github.com/koalaman/shellcheck/wiki/SC2268

Issue: #253