Add automatic certificate renewal for calico/node
Previously, the certificate has been regenerated on every run.
Regenerating the certificate required to restart the calico/node Pods,
which is disruptive to the k8s cluster.

This commit adjusts the generation and setup of calico/node such that the
certificate is only regenerated if it is short to expire or the Typha CA
certificate has changed. In addition, it is checked that the locally
stored calico/node certificate matches the one stored in the Secret.

Issue: #254