Add automatic certificate renewal for Typha
Previously, the certificate has been regenerated on every run.
Regenerating the certificate required to restart the Typha Pods, which
is disruptive to the k8s cluster.

This commit adjusts the generation and setup of Typha such that the
certificate is only regenerated if it is short to expire or the Typha CA
certificate has changed. In addition, it is checked that the locally
stored Typha certificate matches the one stored in the Secret.

Issue: #254