Fix BGP setup for calico
With calico we do disable node-to-node mesh, because it does not scale.
We do setup the masters as route reflectors. All k8s nodes peer with
the masters. Additionally, the masters do peer with the gateways.
From this follows, that a worker peers with all masters. A gateway peers
with all masters. A master is peered with every node.

When the Pod mode is used with ch-k8s-lbaas and an external machine
tries to access an exposed service, the request is received by (the
currently active) gateway. The gateway thinks that the Pod can be reached
via a master node, which is not true, because the masters do not act
as routers and do not forward traffic, they just readvertise BGP routes.
The master will swallow the packets, the Service is unreachable from the
outside.

This commit adjusts the BGP setup such that the masters advertise the
original next hop. The gateway will then forward to the correct k8s node,
instead of trying to use a master as intermediate.

Issue: #227