tf: Improve root disk handling
This commit makes the "block volume as root disk" extension to mk8s a
bit more useful.

1.) Configure the root disk size (if the information is not included in
  the flavor specification).

2.) Set AZ for root disk block volume:

The block volume for the root disk and the instance must be in the same
AZ, otherwise OpenStack will refuse the request. Note that this will
only work if the AZs are explicitly set. I don't care enough right now
to implement this properly because in the usecase of this MR, the AZs
are set.