ipsec: look-up psk from passwordstore
This commit moves the IPsec PSK from ansible vault to passwordstore.
Ansible vault is less flexible from our point of view and should
therefore be avoided.

Note: There's a bootstrapping issue. If the cluster is created from
scratch and IPsec should be configured in the same roll-out, there's a
conflict because the passwordstore hasn't been configured yet.