Remove version from labels
The labels are immutable and this prevents us from doing an easy upgrade.
If we would want to keep the version label, we would need to delete the
daemonsets and deployments before upgrading. This is probably way more
disruptive than the k8s managed rollout of new Pods when a new image
version is used.

Issue: #234