Compare deployed and configured calico version
Check if calico is already deployed on the cluster and if the configured
calico version is unequal to the used one. If that is the case, the
follow up tasks will upgrade the calico components. This is probably
restrictive to the cluster (but I don't think it's disruptive), therefore
a task has been added that fails if disruption is not allowed.

The comparison needs to be in a block so that we can check if the
calicoctl invocation was successful before Ansible tries to resolve
the variables.

While it is theoretically possible to downgrade calico, we do not want
to cover automated downgrades because this may have unpredictable side
effects. Most probably a downgrade is configured by accident anyway.

Issue: #234