VRRP: make priority config more implicit
This commit allows the user to omit the explicit configuration of VRRP
priority for the gateway nodes. If there's ever the need to adapt one of
those values, they can be overwritten via the `vrrp_priorities` variable
(mind the plural!). Note that we rely on the fact that the order of
gateway VMs in the inventory does not change (easily), otherwise a rollout
might be disruptive because of a VIP switch-over.