make lb_ports optional
In ancient times the only way to expose services to the internet was to
configure HAProxy to forward certain ports to NodePorts inside the
cluster. lb_ports describes this mapping. The rise of ch-k8s-lbaas has
made the HAProxy-based load-balancing partially obsolete so there's no
need to keep it enabled by default.