cinder: enable Topology feature gate
With this feature enabled, the controllerplugin respects topological
information like AZs. A PV that's created for a PVC, e.g., would belong
to the same AZ as the node on which the pod runs to which the PV will be
attached. This becomes important when the infrastructure does not allow
cross AZ attachments.