Add role to create wg-based site-to-site (s2s) tunnel
wireguard-site-to-site is a role that allows the user to create
a site-to-site tunnel between two mk8s clusters. Optionally the
pod and/or service subnets can be exposed s.t. one can access a
pod directly "on the other side" without an additional service
involved. Docs can be found under docs/admin/wg.md.