Add devsec ssh hardening
This commit adds a task to apply devsec ssh hardening to all gateway and
k8s hosts. For that, it adds our mirror of the hardening collection as
submodule.

The list of acceptable ciphers etc. was copied from
https://gitlab.cloudandheat.com/cookbooks/ch-ssh-cookbook

It might me a good idea to have a single source of truth, tough.

Fixes #237 #275