Shorten CI job names, rename test stage
The names are too long and have the same prefix. Gitlab shortens the job
names which makes them indistinguishable in the UI. Hovering over the
job names is annoying.

The new naming scheme uses no prefix. Instead, the test stage has been
renamed to "cluster-spawn" to indicate that real resources are created
in these jobs. The jobs are named according to their CNI plugin
and if rook is enabled.

Issue: #264