Configure vm.max_map_count on k8s-workers
Some applications like ElasticSearch require this parameter to be
increased from its default.

I'm not entirely happy with creating a dedicated role for it because the
change could also be achieved from the "user land" by deploying a
privileged pod. On the other hand it saves me the pain of having to
think of it when I replace any of the workers so it's worth it for now.