Create and configure VIPv6 for k8s-API
This commit makes several changes to lay the foundation for a DualStack
based k8s API. However, initializing the k8s control plane with two IP
addresses is currently not possible. The value in "controlPlaneEndpoint"
has to be either ONE IP address or a domain name. Due to the DNS
resolution overhead when using domain name, we agreed on letting the k8s
control plane run on single stack for now. However, this commit makes
all the adjustments to make the k8s control plane use DualStack as soon
as controlPlaneEndpoint may support two IP addresses in the future.

Terraform:
* Adjustments to create an additional fixed IPv6 for the k8s-API
keepalived:
* Adjustments to create an addition VRRP instance for the new VIPv6
* As we now have 3 VRRP instances, group them together
HAproxy:
* bind the VIPv6 to the k8s frontends and backends

It is possible to talk to the k8s API via the VIPv6, because HAproxy
forwards the traffic to the single stack (IPv4) control plane endpoint.

Issue: #227