Refactor k8s cluster bootstrapping / add DualStack
This commit refactors the k8s cluster bootstrapping with kubeadm and
adds configurations and adjustments for the DualStack-support.

Values for the IPv6 Service and Pod network have been added to the
k8s-config role. They are necessary for the DualStack-support and are
used in the kubeadm cluster initialization configuration file.

The k8s cluster initialization config file has been refactored.
This is necessary to tell kubelet which node-IPs it should use from the
start. Additionally, conditional values for the DualStack have been
added to the cluster configuration.

NOTE: DualStack-support is enabled by default starting from k8s v1.21.

The k8s node join procedure has been refactored to use a config file.
It is not possible to mix CLI flags with a configuration file.
Additionally, it is not possible to define the node IPs via the CLI flag.
Therefore, we need to use kubeadm join configuration file only. Using a
configuration file is the preferred way of kubeadm anyway.
Instead of using the "print join command" mechanism of kubeadm, we do
just generate a join token and store it locally. The downside of this is,
that we need to compute the CA hash manually, but this is easy with the
available crypto libraries.

The kubelet flag "cloud-provider" has been removed, because it
interfered with the cluster bootstrapping and led to the collapse of the
kubelet daemon. It is not needed anymore and marked asdeprecated anyway.
The cloud controller manager takes care of it now. Kubelet will remove
the related code from its source in the near future.

Issue: #227