tests: give storage checks more time
A lot of CI runs failed because of, e.g., a CephFS volume that couldn't
be attached. This is a long shot that works on the assumption that rook
simply needs more time to start properly.

The adjusted delay and retry values in check-* were chosen as a
compromise to avoid excessive polling (spams the logs) but also do not
wait longer than necessary.