Add reboot method for cloud-init
Ansible refuses to run the `reboot` task outright when using a
connection type `local` (something about biting hands feeding you).
During initial bootstrap of bare metal clusters, we still need to
reboot.

To do that, we use `command` with `reboot now` to force a reboot,
but only if we are on a local connection. Otherwise, remote ansible
runs will fail with a non-zero exit code (because of the unexpected
loss of connection) and they will not resume after the reboot.

(Note that on bare metal, we use a systemd service which runs at
each boot in order to resume deployment.)